Enables templates for probing in ipv6.
When templates added ip addresses the code still treated them as ipv4 addresses so I updated to handle ipv4 formatting.

The make_packet code in send.c lost the pointer for AES random context. The pointer was overwritten to point to the ip6 src/dst instead. This meant that any time a random number was attempted to be generated we hit a segmentation fault.  To fix this I have updated to create two methods make_packet_ipv6 and make_packet. This allows us to pass the src and dst variables as their own field, and leave the probe_data pointer alone.